feat: Lv2 환전 요청 수행 추가
환전 대상 통화 식별자를 받아 Currency 테이블에 가지고 있는 환율을 기준으로 환전을 수행하도록 구현하였습니다.

+++ src/main/java/com/sparta/currency_user/entity/UserCurrency.java
@@ -3,6 +3,8 @@ package com.sparta.currency_user.entity;
 import jakarta.persistence.*;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Table(name = "user_currency")
@@ -20,13 +22,13 @@ public class UserCurrency extends BaseEntity {
     @JoinColumn(name = "to_currency_id")
     private Currency currency;
 
-    private Long amountInKrw;
+    private BigDecimal amountInKrw;
 
-    private Long amountAfterExchange;
+    private BigDecimal amountAfterExchange;
 
     private String status;
 
-    public UserCurrency(User user, Currency currency, Long amountInKrw, Long amountAfterExchange, String status) {
+    public UserCurrency(User user, Currency currency, BigDecimal amountInKrw, BigDecimal amountAfterExchange, String status) {
         this.user = user;
         this.currency = currency;
         this.amountInKrw = amountInKrw;
